Add `close_connection` function to Conn trait and WebServer impl
A new function named `close_connection` has been added to the `Conn` trait and its implementation in `WebServer`. This function handles the closure of a WebSocket connection by sending a close message to the client and subsequently closing the underlying WebSocket stream. It's called within the `handle_connection` function after echoing a message to the client.