Update readme to include Docker setup instructions
Added instructions in the readme file to guide users on how to set up and run the project using Docker. The updated readme now provides two sets of guidelines: running the project with Docker and without Docker. Docker setup includes steps to build the Docker image and to start the Docker service.